pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["p.minenkov <minikspb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
psycopg2-binary = "^2.9.10"
redis = {extras = ["hiredis"], version = "^5.2.1"}
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
